refactor: Move Google search job scheduling to startWorkers
