+++ src/index.ts
@@ -8,7 +8,6 @@ import openAPIConfig from './config/openapi'
 import { createServerCache } from './createCache'
 import redis from './config/redis'
 import { QUEUE_NAMES } from './queues'
-import googleConfig from './config/google'
 
 export const redisCache = createServerCache({ maxAge: 24 * 60 * 60 * 1000 })
 
@@ -27,72 +26,6 @@ const port = apiConfig.port
 const prisma = new PrismaClient()
 const app = await startApp()
 
-// Schedule daily Google search for new sustainability reports
-async function scheduleGoogleSearchJob() {
-  // Only schedule if Google API is configured
-  if (!googleConfig.isConfigured) {
-    app.log.warn(
-      'Google API not configured. Skipping PDF search job scheduling.'
-    )
-    return
-  }
-
-  const queue = new Queue(QUEUE_NAMES.GOOGLE_SEARCH_PDFS, { connection: redis })
-
-  // Check if job already exists
-  const repeatableJobs = await queue.getRepeatableJobs()
-  const jobExists = repeatableJobs.some(
-    (job) => job.name === 'daily-sustainability-report-search'
-  )
-
-  if (!jobExists) {
-    // Add recurring job to run daily at 2 AM with a consistent job ID
-    await queue.add(
-      'daily-sustainability-report-search',
-      {
-        searchQuery: 'hållbarhetsrapport 2024 filetype:pdf',
-        jobId: 'daily-sustainability-report-search',
-      },
-      {
-        jobId: 'daily-sustainability-report-search',
-        repeat: {
-          pattern: '0 2 * * *', // Cron pattern: At 02:00 every day
-        },
-      }
-    )
-
-    app.log.info('✅ Scheduled daily Google search for sustainability reports')
-  } else {
-    app.log.info('✅ Daily Google search job already scheduled')
-  }
-
-  // Add a one-time test job if requested
-  if (process.env.RUN_SEARCH_TEST === 'true') {
-    // Check if test job is already in the queue
-    const jobs = await queue.getJobs(['waiting', 'active', 'delayed'])
-    const testJobExists = jobs.some(
-      (job) => job.name === 'test-sustainability-report-search'
-    )
-
-    if (!testJobExists) {
-      await queue.add(
-        'test-sustainability-report-search',
-        {
-          searchQuery: 'hållbarhetsrapport 2024 filetype:pdf',
-          threadId: process.env.TEST_THREAD_ID || '0',
-          channelId: process.env.TEST_CHANNEL_ID || '0',
-          jobId: 'test-sustainability-report-search',
-        },
-        {
-          jobId: 'test-sustainability-report-search',
-        }
-      )
-      app.log.info('✅ Added test Google search job')
-    } else {
-      app.log.info('✅ Test Google search job already in queue')
-    }
-  }
-}
 
 async function main() {
   try {
@@ -106,11 +39,6 @@ async function main() {
     }
 
     await app.ready()
-
-    // Schedule the Google search job
-    if (START_BOARD) {
-      await scheduleGoogleSearchJob()
-    }
 
     app.listen(
       {
